moves shared routing-table methods into generated code
* Generates common methods for Table, Fast, and liteTable from one template.
* Maps each table type to its matching node type during generation.
* Moves the public Lite wrappers into litewrapper.go.
